ci: Run tests in parallel (#39988)
Mostly for the coverage tests, let's see if we can speed things up a
bit this way. Looks like it cut the time about in half. 👍

To make this work, we had to generate a separate wp-tests-config file
for each plugin with a different `DB_NAME` in each one so the tests from
different plugins wouldn't deadlock with each other. That then required
an environment variable, `WP_TESTS_CONFIG_FILE_PATH`, to tell the
plugins' test boostrap which config file to use.