coverage: Add missing files/dirs to phpunit configs (#40015)
Ideally everything would be under `src/`, but that's not the case for a
bunch of our packages and plugins.

Also annoying is that PHPUnit doesn't allow globs in `<file>`, so we
have to list any root-level files individually.

There are still some root-level files not touched here that might be,
e.g. `actions.php` files in a few packages or most plugins' entry point
files. Not sure if those should be included or not.

Also, it turns out an earlier PR broke wpcomsh tests. Fix them, and make
lack of `WP_TESTS_CONFIG_FILE_PATH` actually fail the tests.