Migrate to WordPress.com: remove the plugin (#40166)
* Migrate to WordPress.com: remove the plugin

The plugin is now developed by a partner, and no longer by a12s. The version currently available in the Jetpack monorepo no longer matches the version available in the WordPress.org plugin directory.

This commit removes all contents of the plugin in the monorepo, and only leaves indicators that this specific version (and its mirror repo and Packagist package) are no longer maintained, and archived.

* Add some basic files back

* Add lock file back

* Add gitattributes back

* Add more files back

* Add changelog back

* Update lock file

* Remove note from changelog file

I believe it's tripping our changelog validity check

* Remove the plugin entirely

* Remove changelog entry