Repo Gardening: log failures to send Slack messages (#40369)
* Repo Gardening: log failures to send Slack messages

Fixes #40366

Until now, we used to try to send messages to Slack, and didn't do anything if the message could not be sent. We'll now log the message, either by sending a Slack message to myself, or the Quality team for folks outside of the Automattic organization.

This will allow us to better detect triage Slack actions that could not be completed, because a team no longer exists and their Slack channel has been archived for example.

This also changes the Slack messaging utility to use Slack's own package instead of fetch, to give us more access to Slack's error management tools.

* Remove testing config

* Remove extra space